Break NestedBatchUpdate out into it’s own file
